refactor(client): pages folder structure
