--- client/src/pages/Testing.tsx
@@ -1,7 +1,0 @@
-export const Testing = (): JSX.Element => {
-  return (
-    <div className="p-10">
-      <h1 data-testid="testing-title" className="text-3xl font-semibold">Testing Page</h1>
-    </div>
-  );
-};
